+++ app/src/main/kotlin/ac/mdiq/podcini/storage/utils/ChapterUtils.kt
@@ -1,27 +1,16 @@
 package ac.mdiq.podcini.storage.utils
 
-import android.content.ContentResolver
-import android.content.Context
-import android.net.Uri
-import android.util.Log
 import ac.mdiq.podcini.net.download.service.PodciniHttpClient.getHttpClient
 import ac.mdiq.podcini.storage.model.Chapter
-
-import ac.mdiq.podcini.net.feed.parser.media.id3.ChapterReader
-import ac.mdiq.podcini.net.feed.parser.media.id3.ID3ReaderException
-import ac.mdiq.podcini.net.feed.parser.media.vorbis.VorbisCommentChapterReader
-import ac.mdiq.podcini.net.feed.parser.media.vorbis.VorbisCommentReaderException
-import ac.mdiq.podcini.storage.model.Episode
 import ac.mdiq.podcini.util.Logd
+import android.util.Log
 import okhttp3.CacheControl
 import okhttp3.Request
 import okhttp3.Request.Builder
 import okhttp3.Response
-import org.apache.commons.io.input.CountingInputStream
 import org.json.JSONException
 import org.json.JSONObject
-import java.io.*
-import java.util.*
+import java.io.IOException
 import kotlin.math.abs
 
 /**
@@ -29,73 +18,6 @@ import kotlin.math.abs
  */
 object ChapterUtils {
     private val TAG: String = ChapterUtils::class.simpleName ?: "Anonymous"
-
-    @JvmStatic
-    fun getCurrentChapterIndex(media: Episode?, position: Int): Int {
-        val chapters = media?.chapters
-        if (chapters.isNullOrEmpty()) return -1
-        for (i in chapters.indices) if (chapters[i].start > position) return i - 1
-        return chapters.size - 1
-    }
-
-    @JvmStatic
-    fun loadChapters(playable: Episode, context: Context, forceRefresh: Boolean) {
-        // Already loaded
-        if (!forceRefresh) return
-        var chaptersFromDatabase: List<Chapter>? = null
-        var chaptersFromPodcastIndex: List<Chapter>? = null
-        val item = playable
-        if (item.chapters.isNotEmpty()) chaptersFromDatabase = item.chapters
-        if (!item.podcastIndexChapterUrl.isNullOrEmpty()) chaptersFromPodcastIndex = loadChaptersFromUrl(item.podcastIndexChapterUrl!!, forceRefresh)
-
-        val chaptersFromMediaFile = loadChaptersFromMediaFile(playable, context)
-        val chaptersMergePhase1 = merge(chaptersFromDatabase, chaptersFromMediaFile)
-        val chapters = merge(chaptersMergePhase1, chaptersFromPodcastIndex)
-        Logd(TAG, "loadChapters chapters size: ${chapters?.size?:0} ${playable.getEpisodeTitle()}")
-        if (chapters == null) playable.setChapters(listOf())    // Do not try loading again. There are no chapters.
-        else playable.setChapters(chapters)
-    }
-
-    fun loadChaptersFromMediaFile(playable: Episode, context: Context): List<Chapter> {
-        try {
-            openStream(playable, context).use { inVal ->
-                val chapters = readId3ChaptersFrom(inVal)
-                if (chapters.isNotEmpty()) return chapters
-            }
-        } catch (e: IOException) { Log.e(TAG, "Unable to load ID3 chapters: " + e.message)
-        } catch (e: ID3ReaderException) { Log.e(TAG, "Unable to load ID3 chapters: " + e.message) }
-
-        try {
-            openStream(playable, context).use { inVal ->
-                val chapters = readOggChaptersFromInputStream(inVal)
-                if (chapters.isNotEmpty()) return chapters
-            }
-        } catch (e: IOException) { Log.e(TAG, "Unable to load vorbis chapters: " + e.message)
-        } catch (e: VorbisCommentReaderException) { Log.e(TAG, "Unable to load vorbis chapters: " + e.message) }
-        return listOf()
-    }
-
-    @Throws(IOException::class)
-    private fun openStream(playable: Episode, context: Context): CountingInputStream {
-        if (playable.localFileAvailable()) {
-            if (playable.fileUrl == null) throw IOException("No local url")
-            val source = File(playable.fileUrl ?: "")
-            if (!source.exists()) throw IOException("Local file does not exist")
-            return CountingInputStream(BufferedInputStream(FileInputStream(source)))
-        } else {
-            val streamurl = playable.downloadUrl
-            if (streamurl != null && streamurl.startsWith(ContentResolver.SCHEME_CONTENT)) {
-                val uri = Uri.parse(streamurl)
-                return CountingInputStream(BufferedInputStream(context.contentResolver.openInputStream(uri)))
-            } else {
-                if (streamurl.isNullOrEmpty()) throw IOException("stream url is null of empty")
-                val request: Request = Builder().url(streamurl).build()
-                val response = getHttpClient().newCall(request).execute()
-                if (response.body == null) throw IOException("Body is null")
-                return CountingInputStream(BufferedInputStream(response.body!!.byteStream()))
-            }
-        }
-    }
 
     fun loadChaptersFromUrl(url: String, forceRefresh: Boolean): List<Chapter> {
         if (forceRefresh) return loadChaptersFromUrl(url, CacheControl.FORCE_NETWORK)
@@ -114,47 +36,6 @@ object ChapterUtils {
         } catch (e: IOException) { e.printStackTrace()
         } finally { response?.close() }
         return listOf()
-    }
-
-    @Throws(IOException::class, ID3ReaderException::class)
-    private fun readId3ChaptersFrom(inVal: CountingInputStream): List<Chapter> {
-        val reader = ChapterReader(inVal)
-        reader.readInputStream()
-        var chapters = reader.getChapters()
-        chapters = chapters.sortedWith(ChapterStartTimeComparator())
-        enumerateEmptyChapterTitles(chapters)
-        if (!chaptersValid(chapters)) {
-            Logd(TAG, "Chapter data was invalid")
-            return emptyList()
-        }
-        return chapters
-    }
-
-    @Throws(VorbisCommentReaderException::class)
-    private fun readOggChaptersFromInputStream(input: InputStream): List<Chapter> {
-        val reader = VorbisCommentChapterReader(BufferedInputStream(input))
-        reader.readInputStream()
-        var chapters = reader.getChapters()
-        chapters = chapters.sortedWith(ChapterStartTimeComparator())
-        enumerateEmptyChapterTitles(chapters)
-        if (chaptersValid(chapters)) return chapters
-        return emptyList()
-    }
-
-    /**
-     * Makes sure that chapter does a title and an item attribute.
-     */
-    private fun enumerateEmptyChapterTitles(chapters: List<Chapter>) {
-        for (i in chapters.indices) {
-            val c = chapters[i]
-            if (c.title == null) c.title = i.toString()
-        }
-    }
-
-    private fun chaptersValid(chapters: List<Chapter>): Boolean {
-        if (chapters.isEmpty()) return false
-        for (c in chapters) if (c.start < 0) return false
-        return true
     }
 
     /**