+++ app/src/main/kotlin/ac/mdiq/podcini/storage/utils/DurationConverter.kt
@@ -100,7 +100,7 @@ object DurationConverter {
      * @return "HH:MM hours"
      */
     @JvmStatic
-    fun shortLocalizedDuration(context: Context, time: Long, showHoursText: Boolean = true): String {
+    fun durationInHours(context: Context, time: Long, showHoursText: Boolean = true): String {
         val hours = time.toFloat() / 3600f
         return String.format(Locale.getDefault(), "%.2f ", hours) + if (showHoursText) context.getString(R.string.time_hours) else ""
     }